ImageProccesing Results, Json exporter, single File proccesing
Added classes wich holds the data that gets returned
when training or testing the network.
this holds ammount of guesses, averagecost
all image data : image number, image in byte form,
output of neurons, cost of each neuron and number
guessed,

added a class wich holds all data required for
exporting a instance of the network to a .json
this is not yet implemented.

Added the ability to procces a single file.