ImageVerification
Developed a way to create a network with the purpose of
returning a boolean if the specific image
is that networks number.
Each network has its predetermined number.
Is intended to be used for Verifying the results
of the regular network to increase accuracy.

with the thought procces that the Verifyer
will have a higher accuracy in a boolean
then a question with 10 answers.

There are still some problems with this,
for instance the deserialization still happens in a
very clumsy way, weather or not this actually has
a affect on accuracy is still to be determined,
there still seem to be some issues with serialization.